Moves FormFieldImageUpload
